# aspnet-openapi-examples.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "aspnet-openapi-examples | v1",
    "description": "# Examples of using ASP.NET to create OpenAPI v3 API definitions\n\nThis is an example of the info.description content. This description and most other\ndescription fields in OpenAPI v3.x support [CommonMark] markdown content.\n\n[CommonMark]: https://spec.commonmark.org/\n\nBelow are examples of the CommonMark markdown elements.\n\n## Headings\n\n### H3\n\n## Text styling\n\n**bold text**\n\n*italicized text*\n\n> blockquote\n\n`code`\n\n## Lists\n\n### Ordered List\n\n1. First item\n2. Second item\n3. Third item\n\n### Unordered List\n\n- First item\n- Second item\n- Third item\n\n## Horizontal Rule\n\n---\n\n### Links\n\n[Markdown Guide](https://www.markdownguide.org)\n\n### Image\n\n![alt text](https://www.markdownguide.org/assets/images/tux.png)\n\n### Table\n\n| Syntax | Description |\n| ----------- | ----------- |\n| Header | Title |\n| Paragraph | Text |\n\n### Fenced Code Block\n\n```\n{\n\"firstName\": \"John\",\n\"lastName\": \"Smith\",\n\"age\": 25\n}\n```\n",
    "contact": {
      "name": "Contoso Admin",
      "url": "https://contoso.com/contacts",
      "email": "admin@contoso.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0",
    "x-my-extension": "my-extension-value"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local development server"
...
  "paths": {
    "/request-bodies/no-consumes": {
      "post": {
        "tags": [
          "RequestBodies"
        ],
        "requestBody": {
          "description": "The request body",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Body"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Body"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Body"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/Body"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/request-bodies/consumes-json": {
      "post": {
        "tags": [
          "RequestBodies"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/request-bodies/bad-consumes": {
      "post": {
        "tags": [
          "RequestBodies"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/request-bodies/xml-body": {
      "post": {
        "tags": [
          "RequestBodies"
        ],
        "requestBody": {
          "description": "An Xml request body",
          "content": {
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/XmlBody"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/XmlBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/request-bodies/text-body": {
      "post": {
        "tags": [
          "RequestBodies"
        ],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/request-bodies/stream-body": {
      "post": {
        "tags": [
          "RequestBodies"
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "$ref": "#/components/schemas/Stream"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body": {
        "type": "object",
        "properties": {
          "prop1": {
            "type": "string"
          },
          "prop2": {
            "type": "string",
            "nullable": true
          },
          "prop3": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Stream": {
        "type": "string",
        "format": "binary"
      },
      "XmlBody": {
        "type": "object",
        "properties": {
          "prop1": {
            "type": "string"
          },
          "prop2": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "prop3": {
            "type": "boolean",
            "nullable": true
          }
        }
      }
    }
  },
...
}